Using Datatables for the ownerships page
Both aldeed:tabular and ephemer:reactive-datatables work well.
Although the first is optimized for server side filtering, and is much more complex, while latter is for client side filtering and is simpler.
I suggest to use the latter, until tables grow so big, that filtering should happen on the server.